Move styles for components /page to separate file
