+++ src/components/account/index.js
@@ -1,0 +1,8 @@
+export * from './AccountButtons'
+export * from './AccountDetails'
+export * from './AccountAddress'
+export * from './AccountForm'
+export * from './AccountImage'
+export * from './AccountImageUpload'
+export * from './AccountLogIn'
+export * from './AccountOptions'
