+++ src/components/styles/AccountDetails.styles.js
@@ -1,0 +1,17 @@
+export default {
+    Row: {
+        display: 'flex',
+        flex: 1,
+        flexDirection: 'row',
+    },
+    Column: {
+        flexDirection: 'column',
+        justifyContent: 'center'
+    },
+    RowBtn: {
+        display: 'flex',
+        flex: 1,
+        flexDirection: 'row',
+        paddingBottom: 15
+    } 
+}
