+++ src/components/styles/AccountForm.styles.js
@@ -1,0 +1,19 @@
+export default {
+    scrollView: {
+        marginHorizontal: 20,
+      },
+      imgSty: {
+          justifyContent: 'center',
+          alignItems: 'center',
+          padding: 50
+      },
+      Column: {
+          display: 'flex',
+          flexDirection: 'column',
+          flexBasis: 90,
+          justifyContent: 'center',
+          alignItems: 'center',
+          flex: 1,
+          margin: 10
+        }
+}
