+++ src/components/styles/BottomBtn.styles.js
@@ -1,0 +1,16 @@
+export default {
+    btnBtmStyle: {
+        textColor: 'black',
+        flex: 1
+    },
+    txtBtnStyle: {
+        color: 'gray'
+    },
+    bottomView:{
+        paddingUp: 10,
+        alignItems: 'center',
+        borderTopColor: 'color',
+        borderTopWidth: 2,
+        marginBottom: 20,
+      }
+}
