+++ src/components/styles/ErrorText.styles.js
@@ -1,0 +1,9 @@
+export default {
+    errorTextStyle: {
+        fontSize: 20,
+        alignSelf: 'center',
+        color: 'red',
+        backgroundColor:  '#ffffff',
+        justifyContent: 'center'
+      }
+}
