+++ src/components/styles/General.styles.js
@@ -1,0 +1,11 @@
+export default {
+    Row: {
+        display: 'flex',
+        flex: 1,
+        flexDirection: 'row',
+    },
+    Column: {
+        flexDirection: 'column',
+        justifyContent: 'center'
+    }
+}
