+++ src/components/styles/LogInBtn.styles.js
@@ -1,0 +1,24 @@
+export default {
+    textStyle: {
+      alignSelf: 'center',
+      color: '#fff',
+      fontSize: 16,
+      fontWeight: '600',
+      paddingTop: 5,
+      paddingBottom: 5,        
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    buttonStyle: {
+      flex: 1,
+      backgroundColor:'#4FA6FD' ,
+      borderRadius: 5,
+      borderWidth: 1,
+      borderColor: '#fff',
+      marginLeft: 10,
+      marginRight: 10,
+      width: 250, 
+      justifyContent: 'center',
+      alignItems: 'center',
+    }
+  };
