+++ src/components/styles/SignUpBtn.styles.js
@@ -1,0 +1,24 @@
+export default {
+    textStyle: {
+      alignSelf: 'center',
+      color: '#fff',
+      fontSize: 16,
+      fontWeight: '600',
+      paddingTop: 10,
+      paddingBottom: 10,        
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    buttonStyle: {
+      flex: 1,
+      backgroundColor: '#724FFD' ,
+      borderRadius: 0,
+      borderWidth: 1,
+      borderColor: '#fff',
+      marginLeft: 5,
+      marginRight: 5,
+      width: 250,        
+      justifyContent: 'center',
+      alignItems: 'center'
+    }
+  };
