+++ src/components/styles/index.js
@@ -1,3 +1,9 @@
-export * from './LogInBtn'
-export * from './SignUpBtn'
-export * from './BottomBtn'
+
+export * from './AccountDetails.styles'
+export * from './AccountForm.styles'
+export * from './BottomBtn.styles'
+export * from './ClearBtnSty.styles'
+export * from './DeleteProfileBtn'
+export * from './ErrorText.styles'
+export * from './LogInBtn.styles'
+export * from './SignUpBtn.styles'
