+++ src/navigation/drawer/AppMenu.style.js
@@ -1,0 +1,19 @@
+export default {
+    container: {
+      paddingTop: 20,
+      flex: 1
+    },
+    navItemStyle: {
+      padding: 10
+    },
+    navSectionStyle: {
+      backgroundColor: 'light'
+    },
+    sectionHeadingStyle: {
+      paddingVertical: 10,
+      paddingHorizontal: 5
+    },
+    footerContainer: {
+      backgroundColor: 'light'
+    }
+  };
