+++ src/navigation/navButtons/LogOutNav.styles.js
@@ -1,0 +1,10 @@
+export default {
+    btnView: {
+        justifyContent: 'center',
+        alignItems: 'center',
+        paddingRight: 10
+    }, 
+    txtLabel: {
+        paddingRight: 5
+    }
+}
