+++ src/navigation/navButtons/LogoNav.js
@@ -1,14 +1,23 @@
 import React from 'react'
 import { ImageButton } from '../../components/common'
+import { NavigationActions } from 'react-navigation'
 
 /**
  * Logo button for navigation bar
  */
 class LogoNav extends React.Component {
+    navigateToScreen = (route) => () => {
+        const navigateAction = NavigationActions.navigate({
+            routeName: route
+        });
+
+        this.props.navigation.dispatch(navigateAction)
+    }
+
     render(){
         return (
             <ImageButton
-                onPress={() => this.props.navigation.navigate('Home')}
+                onPress={() => this.navigateToScreen('Home')}
                 imageSource={require('../../image/Placeholder150.png')}
             />
         )
