+++ src/navigation/stack/AppNavigator.js
@@ -1,14 +1,10 @@
 
 import React from 'react'
-
-// import { createStackNavigator  } from 'react-navigation-stack'
-import { createDrawerNavigator } from 'react-navigation-drawer'
+import { createDrawerNavigator,  } from 'react-navigation-drawer'
 
 import Profile from '../../page/account/Profile'
 import Home from '../../page/general/Home'
-
-import LogoNav from '../navButtons/LogoNav'
-// import SignOutNav from '../navButtons/SignOutNav'
+import AppMenu from '../drawer/AppMenu'
 
 /**
  * Application menu accessible page
@@ -19,10 +15,8 @@ const AppNavigator = createDrawerNavigator(
         "Profile": Profile
     },
     {
-       initialRouteName: "Profile",
-       defaultNavigationOptions : {
-        headerRight: <LogoNav />
-    }
+        initialRouteName: "Home",
+        contentComponent: AppMenu,
     }
 );
