+++ src/navigation/stack/AuthNavigator.js
@@ -1,5 +1,4 @@
 import React from 'react'
-// import firebase from 'firebase'
 import { createStackNavigator  } from 'react-navigation-stack'
 
 import LogInOptions from '../../page/account/LogInOptions'
@@ -8,10 +7,9 @@ import LogInEmail from '../../page/account/LogInEmail'
 import ForgotPassword from '../../page/account/ForgotPassword'
 import SignUpEmail from '../../page/account/SignUpEmail'
 
-import AccountLogOut from '../../components/account/AccountLogOut'
+// import AccountLogOut from '../../components/account/AccountLogOut'
 
 import LogoNav from '../navButtons/LogoNav'
-// import SignOutNav from '../navButtons/SignOutNav'
 
 /**
  * Auth menu accessible page
@@ -23,12 +21,11 @@ const AuthNavigator = createStackNavigator(
         "SignUp": SignUpOptions,
         "Sign Up With Email": SignUpEmail,
         "Login": LogInOptions,
-        "LogOut": AccountLogOut
     }, {
         initialRouteName: "Login",       
-        defaultNavigationOptions : {
-            headerRight: <LogoNav />,
-        }
+        // defaultNavigationOptions : {
+        //     headerRight: <LogoNav />,
+        // }
     }
 );
