+++ src/page/account/SignUpOptions.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import { View, ScrollView } from 'react-native'
-import LogInBtn from '../../components/styles/LogInBtn'
-import BottomBtn from '../../components/styles/BottomBtn'
+import styles from '../styles/SignUpOptions.styles'
+import LogInBtn from '../../components/styles/LogInBtn.styles'
+import BottomBtn from '../../components/styles/BottomBtn.styles'
 import AccountImage from '../../components/account/AccountImage'
 import AccountButtons from '../../components/account/AccountButtons'
 import AccountOptions from '../../components/account/AccountOptions'
@@ -39,31 +40,6 @@ class SignUpOptions extends React.Component {
             </ScrollView>
         )
     }
-}
-
-const styles = {
-  imgSty: {
-      justifyContent: 'center',
-      alignItems: 'center',
-      padding: 50
-  },
-  scrollView: {
-    marginHorizontal: 20,
-  },
-  Row: {
-      display: 'flex',
-      flexDirection: 'row',
-    },
-  Column: {
-      display: 'flex',
-      flexDirection: 'column',
-      flexBasis: 90,
-      justifyContent: 'center',
-      alignItems: 'center',
-      flex: 1,
-      margin: 10
-    }
-
 }
 
 export default SignUpOptions
