+++ src/page/account/SignUpStuff.js
@@ -1,1 +1,23 @@
+import React from 'react'
+import { Text } from 'react-native'
+
+class SignUpStuff extends React.Component {
+    render(){
+        return (
+            <Text>{'SignUp Stuff'}</Text>
+        )
+    }
+}
+
+export default SignUpStuff
 //https://reactnavigation.org/docs/en/deep-linking.html
+
+// Add form validation on all auth Pages
+// - Sign-Up
+// - Log-In
+// - Profile
+
+// Minor Route changes
+// - Changed from stack to draw after Logging in.
+
+// - Updated Front to handle gateway Autherican
