+++ src/page/general/Home.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { Text } from 'react-native'
+import firbase from 'firebase'
+import { Text, View } from 'react-native'
 
 /**
  * Home Page
@@ -7,8 +8,18 @@ import { Text } from 'react-native'
 class Home extends React.Component {
     render(){
         return (
-            <Text>{'Home'}</Text>
+            <View>
+                <Text style={styles.textStyles}>{'Home'}</Text>
+                <Text style={styles.textStyles}>{firbase.auth().currentUser.email}</Text>
+                <Text style={styles.textStyles}>{firbase.auth().currentUser.displayName}</Text>
+            </View>
         )
+    }
+}
+
+const styles = {
+    textStyles: {
+        textAlign: 'center',      
     }
 }
