+++ src/page/styles/Loader.styles.js
@@ -1,0 +1,7 @@
+export default {
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center'
+      }
+}
