+++ src/page/styles/LogInOptions.styles.js
@@ -1,0 +1,23 @@
+export default {
+    imgSty: {
+        justifyContent: 'center',
+        alignItems: 'center',
+        padding: 50
+    },
+    scrollView: {
+      marginHorizontal: 20,
+    },
+    Row: {
+        display: 'flex',
+        flexDirection: 'row',
+      },
+    Column: {
+        display: 'flex',
+        flexDirection: 'column',
+        flexBasis: 90,
+        justifyContent: 'center',
+        alignItems: 'center',
+        flex: 1,
+        margin: 10
+      }
+}
