+++ src/page/styles/Profile.styles.js
@@ -1,0 +1,21 @@
+export default {
+    Row: {
+        display: 'flex',
+        flex: 1,
+        flexDirection: 'row',
+    },
+    Column: {
+        display: 'flex',
+        flex: 1,
+        flexDirection: 'column',
+        justifyContent: 'space-between'
+    },
+    scrollView: {
+    //   marginHorizontal: 20,
+      backgroundColor:  '#ffffff'
+    },
+    View: {
+      justifyContent: 'center',
+      backgroundColor:  '#ffffff'
+    }
+}
