+++ src/reducer/auth.js
@@ -5,11 +5,10 @@ export default (state = null, action ) => {
     let { type, payload } = action;
 
     switch (type) {
-        case 'USER_SET': {
-            console.log(payload);
-            
-            return payload
-        }
+        case 'USER_SET': 
+            return payload;
+        case 'TOKEN': 
+            return payload;
         case 'AUTH_ERROR': return payload
         default: return state;
     }
